+++ rmtest/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility package exposing calibration modules under ``rmtest``."""
