+++ rmtest/calibration.py
@@ -1,0 +1,9 @@
+import importlib.util
+import sys
+from pathlib import Path
+
+_path = Path(__file__).resolve().parent.parent / 'calibration.py'
+_spec = importlib.util.spec_from_file_location(__name__, _path)
+_module = importlib.util.module_from_spec(_spec)
+sys.modules[__name__] = _module
+_spec.loader.exec_module(_module)
